mocks: factor our status handler mock
