+++ src/mock/mod.rs
@@ -1,10 +1,10 @@
 #![cfg(test)]
 
-use crate::*;
+// use crate::*;
+use crate::{Module, Trait};
 
 use primitives::{Blake2Hasher, H256};
 
-use crate::{Module, Trait};
 use balances;
 use minting;
 use runtime_primitives::{
@@ -14,74 +14,11 @@ use runtime_primitives::{
 };
 use srml_support::{impl_outer_origin, parameter_types};
 
+mod status_handler;
+pub use status_handler::{MockStatusHandler};
+
 impl_outer_origin! {
     pub enum Origin for Test {}
-}
-use std::cell::RefCell;
-
-struct StatusHandlerState {
-    successes: Vec<RewardRelationshipId>,
-    failures: Vec<RewardRelationshipId>,
-}
-impl StatusHandlerState {
-    pub fn reset(&mut self) {
-        self.successes = vec![];
-        self.failures = vec![];
-    }
-}
-impl Default for StatusHandlerState {
-    fn default() -> Self {
-        Self {
-            successes: vec![],
-            failures: vec![],
-        }
-    }
-}
-
-thread_local!(static STATUS_HANDLER_STATE: RefCell<StatusHandlerState> = RefCell::new(Default::default()));
-
-pub struct MockStatusHandler {}
-impl MockStatusHandler {
-    pub fn reset() {
-        STATUS_HANDLER_STATE.with(|cell| {
-            cell.borrow_mut().reset();
-        });
-    }
-    pub fn successes() -> usize {
-        let mut value = 0;
-        STATUS_HANDLER_STATE.with(|cell| {
-            value = cell.borrow_mut().successes.len();
-        });
-        value
-    }
-    pub fn failures() -> usize {
-        let mut value = 0;
-        STATUS_HANDLER_STATE.with(|cell| {
-            value = cell.borrow_mut().failures.len();
-        });
-        value
-    }
-}
-impl<T: Trait> PayoutStatusHandler<T> for MockStatusHandler {
-    fn payout_succeeded(
-        id: RewardRelationshipId,
-        _destination_account: &T::AccountId,
-        _amount: BalanceOf<T>,
-    ) {
-        STATUS_HANDLER_STATE.with(|cell| {
-            cell.borrow_mut().successes.push(id);
-        });
-    }
-
-    fn payout_failed(
-        id: RewardRelationshipId,
-        _destination_account: &T::AccountId,
-        _amount: BalanceOf<T>,
-    ) {
-        STATUS_HANDLER_STATE.with(|cell| {
-            cell.borrow_mut().failures.push(id);
-        });
-    }
 }
 
 // Workaround for https://github.com/rust-lang/rust/issues/26925 . Remove when sorted.
